[IMP] sale_timesheet: factorize common class tests to be reused
Impacted modules: sale_timesheet and sale_project_timesheet

All tests were moved from sale_timesheet to sale_project_timesheet to avoid
rewrite them all to remove the project dependency. However, some methods of
the common class tests (the ones generating test data) can be usefull for
other module.
This commit proposes a factorization of some part of those methods, in order
to be reuse in other modules (like timesheet_grid_sale).

Task-1923949
