+++ addons/sale_project_timesheet/tests/common.py
@@ -1,33 +1,16 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from odoo.addons.sale.tests.test_sale_common import TestCommonSaleNoChart
+from odoo.addons.sale_timesheet.tests.common import TestCommonSaleTimesheetNoChart
 
 
-class TestCommonSaleTimesheetNoChart(TestCommonSaleNoChart):
-
-    @classmethod
-    def setUpEmployees(cls):
-        # Create employees
-        cls.employee_user = cls.env['hr.employee'].create({
-            'name': 'Employee User',
-            'timesheet_cost': 15,
-        })
-        cls.employee_manager = cls.env['hr.employee'].create({
-            'name': 'Employee Manager',
-            'timesheet_cost': 45,
-        })
+class TestCommonSaleProjectTimesheetNoChart(TestCommonSaleTimesheetNoChart):
 
     @classmethod
     def setUpServiceProducts(cls):
-        """ Create Service product for all kind, with each tracking policy. """
-        # Account and project
-        cls.account_sale = cls.env['account.account'].create({
-            'code': 'SERV-2020',
-            'name': 'Product Sales - (test)',
-            'reconcile': True,
-            'user_type_id': cls.env.ref('account.data_account_type_revenue').id,
-        })
+        """ Create service tracked product, based on project and task's timesheet. """
+
+        super(TestCommonSaleProjectTimesheetNoChart, cls).setUpServiceProducts()
 
         # Create projects
         cls.project_global = cls.env['project.project'].create({
@@ -48,21 +31,7 @@ class TestCommonSaleTimesheetNoChart(TestCommonSaleNoChart):
         uom_hour = cls.env.ref('uom.product_uom_hour')
 
         # -- ordered quantities (ordered, timesheet)
-        cls.product_order_timesheet1 = cls.env['product.product'].create({
-            'name': "Service Ordered, create no task",
-            'standard_price': 11,
-            'list_price': 13,
-            'type': 'service',
-            'invoice_policy': 'order',
-            'uom_id': uom_hour.id,
-            'uom_po_id': uom_hour.id,
-            'default_code': 'SERV-ORDERED1',
-            'service_type': 'timesheet',
-            'service_tracking': 'no',
-            'project_id': False,
-            'taxes_id': False,
-            'property_account_income_id': cls.account_sale.id,
-        })
+        # Note: cls.product_order_timesheet1 is tracked by timesheet but without creating task
         cls.product_order_timesheet2 = cls.env['product.product'].create({
             'name': "Service Ordered, create task in global project",
             'standard_price': 30,
@@ -126,21 +95,7 @@ class TestCommonSaleTimesheetNoChart(TestCommonSaleNoChart):
         })
 
         # -- timesheet on tasks (delivered, timesheet)
-        cls.product_delivery_timesheet1 = cls.env['product.product'].create({
-            'name': "Service delivered, create no task",
-            'standard_price': 11,
-            'list_price': 13,
-            'type': 'service',
-            'invoice_policy': 'delivery',
-            'uom_id': uom_hour.id,
-            'uom_po_id': uom_hour.id,
-            'default_code': 'SERV-DELI1',
-            'service_type': 'timesheet',
-            'service_tracking': 'no',
-            'project_id': False,
-            'taxes_id': False,
-            'property_account_income_id': cls.account_sale.id,
-        })
+        # Note: cls.product_delivery_timesheet1 is tracked by timesheet but without creating task
         cls.product_delivery_timesheet2 = cls.env['product.product'].create({
             'name': "Service delivered, create task in global project",
             'standard_price': 30,
@@ -204,21 +159,7 @@ class TestCommonSaleTimesheetNoChart(TestCommonSaleNoChart):
         })
 
         # -- milestones (delivered, manual)
-        cls.product_delivery_manual1 = cls.env['product.product'].create({
-            'name': "Service delivered, create no task",
-            'standard_price': 11,
-            'list_price': 13,
-            'type': 'service',
-            'invoice_policy': 'delivery',
-            'uom_id': uom_hour.id,
-            'uom_po_id': uom_hour.id,
-            'default_code': 'SERV-DELI1',
-            'service_type': 'manual',
-            'service_tracking': 'no',
-            'project_id': False,
-            'taxes_id': False,
-            'property_account_income_id': cls.account_sale.id,
-        })
+        # Note: cls.product_delivery_manual1 is tracked manual does not create task nor project
         cls.product_delivery_manual2 = cls.env['product.product'].create({
             'name': "Service delivered, create task in global project",
             'standard_price': 30,
