+++ addons/sale_project_timesheet/tests/test_reinvoice.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from odoo.addons.sale_project_timesheet.tests.common import TestCommonSaleTimesheetNoChart
+from odoo.addons.sale_project_timesheet.tests.common import TestCommonSaleProjectTimesheetNoChart
 
 
-class TestReInvoice(TestCommonSaleTimesheetNoChart):
+class TestReInvoice(TestCommonSaleProjectTimesheetNoChart):
 
     @classmethod
     def setUpClass(cls):
