+++ addons/sale_project_timesheet/tests/test_sale_service.py
@@ -1,11 +1,11 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from odoo.addons.sale_project_timesheet.tests.common import TestCommonSaleTimesheetNoChart
+from odoo.addons.sale_project_timesheet.tests.common import TestCommonSaleProjectTimesheetNoChart
 from odoo.exceptions import UserError, ValidationError
 
 
-class TestSaleService(TestCommonSaleTimesheetNoChart):
+class TestSaleService(TestCommonSaleProjectTimesheetNoChart):
     """ This test suite provide checks for miscellaneous small things. """
 
     @classmethod
